Implement lazy matcher that checks for Empty
If the language can be identified as being empty, we can quickly return
right without looking at more of the string or expanding the derivative.

There is still some inefficiency due to re-checking for an empty
language at every character - it would be better to simplify the regex
along the way to make this faster.